+++ EngineYard_iOS/ViewControllers/HUDView.xib
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11762" systemVersion="15G1510" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+    <device id="retina4_0" orientation="landscape">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11757"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="HUDView" customModule="EngineYard" customModuleProvider="target">
+            <connections>
+                <outlet property="collectionView" destination="ZNM-jU-zo3" id="Z7v-98-tHX"/>
+                <outlet property="menuBtn" destination="7Tb-fj-zEf" id="pfN-fw-etO"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="iN0-l3-epB">
+            <rect key="frame" x="0.0" y="0.0" width="568" height="50"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" showsVerticalScrollIndicator="NO" keyboardDismissMode="onDrag" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="ZNM-jU-zo3">
+                    <rect key="frame" x="0.0" y="0.0" width="488" height="50"/>
+                    <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                    <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" minimumLineSpacing="10" minimumInteritemSpacing="10" id="DUB-ga-VQB">
+                        <size key="itemSize" width="100" height="40"/>
+                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
+                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
+                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
+                    </collectionViewFlowLayout>
+                    <connections>
+                        <outlet property="dataSource" destination="-1" id="241-zS-zh2"/>
+                        <outlet property="delegate" destination="-1" id="IS2-0K-6gq"/>
+                    </connections>
+                </collectionView>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="7Tb-fj-zEf">
+                    <rect key="frame" x="507" y="10" width="39" height="30"/>
+                    <state key="normal" title="Menu"/>
+                    <connections>
+                        <action selector="menuBtnPressed:" destination="-1" eventType="touchUpInside" id="pMj-4s-oYn"/>
+                    </connections>
+                </button>
+            </subviews>
+            <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+            <constraints>
+                <constraint firstAttribute="trailing" secondItem="7Tb-fj-zEf" secondAttribute="trailing" constant="22" id="OKr-QI-nCQ"/>
+                <constraint firstItem="7Tb-fj-zEf" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="SSH-Kk-8Ys"/>
+                <constraint firstItem="ZNM-jU-zo3" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="Tws-hQ-1VZ"/>
+                <constraint firstItem="7Tb-fj-zEf" firstAttribute="leading" secondItem="ZNM-jU-zo3" secondAttribute="trailing" constant="19" id="USz-6X-ovd"/>
+                <constraint firstItem="ZNM-jU-zo3" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="Uht-DZ-zRH"/>
+                <constraint firstAttribute="bottom" secondItem="ZNM-jU-zo3" secondAttribute="bottom" id="lTm-7j-aR4"/>
+            </constraints>
+            <nil key="simulatedStatusBarMetrics"/>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+        </view>
+    </objects>
+</document>
