--- EngineYard_iOS/ViewControllers/HUDViewController.xib
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11762" systemVersion="15G1510" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="retina4_0" orientation="landscape">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11757"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="HUDViewController" customModule="EngineYard" customModuleProvider="target">
-            <connections>
-                <outlet property="collectionView" destination="ZSM-2S-UIA" id="5cW-7J-fdk"/>
-                <outlet property="menuButton" destination="zsh-7b-bhn" id="ae3-z8-408"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="568" height="50"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" showsVerticalScrollIndicator="NO" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="ZSM-2S-UIA">
-                    <rect key="frame" x="0.0" y="0.0" width="479" height="50"/>
-                    <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                    <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" minimumLineSpacing="10" minimumInteritemSpacing="10" id="2uV-Fi-sVg">
-                        <size key="itemSize" width="100" height="40"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                    <connections>
-                        <outlet property="dataSource" destination="-1" id="UIT-X1-nTl"/>
-                        <outlet property="delegate" destination="-1" id="82T-T0-G0U"/>
-                    </connections>
-                </collectionView>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="zsh-7b-bhn">
-                    <rect key="frame" x="499" y="10" width="39" height="30"/>
-                    <state key="normal" title="Menu"/>
-                    <connections>
-                        <action selector="menuButtonPressed:" destination="-1" eventType="touchUpInside" id="HTj-vg-nTQ"/>
-                    </connections>
-                </button>
-            </subviews>
-            <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-            <constraints>
-                <constraint firstItem="zsh-7b-bhn" firstAttribute="leading" secondItem="ZSM-2S-UIA" secondAttribute="trailing" constant="20" id="1Eh-Iu-6jq"/>
-                <constraint firstItem="zsh-7b-bhn" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="10" id="7t9-rL-NGM"/>
-                <constraint firstItem="ZSM-2S-UIA" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="8sm-TV-GnY"/>
-                <constraint firstAttribute="bottom" secondItem="zsh-7b-bhn" secondAttribute="bottom" constant="10" id="AJV-Ko-S3Z"/>
-                <constraint firstAttribute="trailing" secondItem="zsh-7b-bhn" secondAttribute="trailing" constant="30" id="GYw-6y-oFj"/>
-                <constraint firstAttribute="bottom" secondItem="ZSM-2S-UIA" secondAttribute="bottom" id="PvK-5k-imC"/>
-                <constraint firstItem="ZSM-2S-UIA" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="jHD-i4-yis"/>
-            </constraints>
-            <nil key="simulatedStatusBarMetrics"/>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-        </view>
-    </objects>
-</document>
