+++ EngineYard_iOS/Views/PlayerCard/PlayerHUDView.xib
